+++ src/sys/windows/mod.rs
@@ -548,7 +548,15 @@ impl Drop for RegKey {
 
 pub fn enumerate() -> std::io::Result<Vec<PathBuf>> {
 	let subkey = unsafe { CStr::from_bytes_with_nul_unchecked(b"Hardware\\DEVICEMAP\\SERIALCOMM\x00") };
-	let device_map = RegKey::open(winreg::HKEY_LOCAL_MACHINE, subkey, winnt::KEY_READ)?;
+	let device_map = match RegKey::open(winreg::HKEY_LOCAL_MACHINE, subkey, winnt::KEY_READ) {
+		Ok(x) => x,
+		Err(ref e) if e.kind() == std::io::ErrorKind::NotFound => {
+			// The registry key doesn't exist until a serial port was available at-least once.
+			return Ok(Vec::new());
+		},
+		Err(e) => return Err(e),
+	};
+
 	let (value_count, max_value_name_len, max_value_data_len) = device_map.get_value_info()?;
 
 	let mut entries = Vec::with_capacity(16);